[feat/#34]: 홈 navigation 루트 변경
