+++ app/src/main/java/com/umc/coumo/presentation/fragment/HomeMenuFragment.kt
@@ -4,16 +4,17 @@ import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.lifecycleScope
+import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.umc.coumo.R
 import com.umc.coumo.databinding.FragmentHomeMenuBinding
 import com.umc.coumo.domain.viewmodel.HomeViewModel
 import com.umc.coumo.presentation.adapter.MenuAdapter
-import com.umc.coumo.utils.binding.BindingFragmentNoneBackPress
+import com.umc.coumo.utils.binding.BindingFragment
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 
-class HomeMenuFragment: BindingFragmentNoneBackPress<FragmentHomeMenuBinding>(R.layout.fragment_home_menu) {
+class HomeMenuFragment: BindingFragment<FragmentHomeMenuBinding>(R.layout.fragment_home_menu) {
 
     private val viewModel : HomeViewModel by activityViewModels ()
 
@@ -21,7 +22,7 @@ class HomeMenuFragment: BindingFragmentNoneBackPress<FragmentHomeMenuBinding>(R.
         super.onViewCreated(view, savedInstanceState)
 
         setRecyclerView()
-
+        setButton()
     }
 
     private fun setRecyclerView() {
@@ -37,5 +38,16 @@ class HomeMenuFragment: BindingFragmentNoneBackPress<FragmentHomeMenuBinding>(R.
                 menuAdapter.submitList(it)
             }
         }
+    }
+
+    private fun setButton() {
+        binding.btnToolbarBack.setOnClickListener {
+            onBackPressed()
+        }
+    }
+
+    override fun onBackPressed() {
+        super.onBackPressed()
+        findNavController().popBackStack()
     }
 }
