+++ app/src/main/res/layout/fragment_home_menu.xml
@@ -6,53 +6,89 @@
 
     </data>
 
-    <androidx.core.widget.NestedScrollView
+    <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
-        android:layout_height="wrap_content">
+        android:layout_height="match_parent">
 
         <androidx.constraintlayout.widget.ConstraintLayout
+            android:id="@+id/section_tool_bar"
             android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:paddingBottom="@dimen/bottom_above">
+            style="@style/TabBarDefaultStyle"
+            app:layout_constraintTop_toTopOf="parent">
 
-            <TextView
-                android:id="@+id/tv_store_name"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginHorizontal="@dimen/horizontal_padding"
-                android:layout_marginTop="24dp"
-                android:text="앙떼띠 로스터리 (강남점)"
-                android:textSize="24sp"
-                android:textColor="@color/font_text"
-                android:fontFamily="@font/pretendard_700"
+            <androidx.constraintlayout.utils.widget.ImageFilterView
+                android:id="@+id/btn_toolbar_back"
+                android:layout_width="30dp"
+                android:layout_height="30dp"
+                android:src="@drawable/icon_chevron_left"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent"/>
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintBottom_toBottomOf="parent"/>
 
             <TextView
-                android:id="@+id/tv_coupon_count"
+                android:id="@+id/tv_toolBar_title"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:layout_marginTop="10dp"
-                android:paddingHorizontal="18dp"
-                android:paddingVertical="6dp"
-                android:background="@drawable/shape_rect_side_round"
-                android:backgroundTint="#E4D9F8"
-                android:text="쿠모로 4개 모았어요!"
-                android:textSize="12sp"
-                android:textColor="#4E4E4E"
+                android:text="카페/디저트"
+                android:textColor="@color/main"
+                android:textSize="18sp"
                 android:fontFamily="@font/pretendard_600"
-                app:layout_constraintStart_toStartOf="@id/tv_store_name"
-                app:layout_constraintTop_toBottomOf="@id/tv_store_name"/>
-
-            <androidx.recyclerview.widget.RecyclerView
-                android:id="@+id/rv_menu"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="40dp"
-                android:paddingHorizontal="@dimen/horizontal_padding"
-                app:layout_constraintTop_toBottomOf="@+id/tv_coupon_count"/>
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintBottom_toBottomOf="parent"/>
 
         </androidx.constraintlayout.widget.ConstraintLayout>
 
-    </androidx.core.widget.NestedScrollView>
+        <androidx.core.widget.NestedScrollView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            app:layout_constraintTop_toBottomOf="@id/section_tool_bar">
+
+            <androidx.constraintlayout.widget.ConstraintLayout
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:paddingBottom="@dimen/bottom_above">
+
+                <TextView
+                    android:id="@+id/tv_store_name"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_marginHorizontal="@dimen/horizontal_padding"
+                    android:layout_marginTop="24dp"
+                    android:text="앙떼띠 로스터리 (강남점)"
+                    android:textSize="24sp"
+                    android:textColor="@color/font_text"
+                    android:fontFamily="@font/pretendard_700"
+                    app:layout_constraintStart_toStartOf="parent"
+                    app:layout_constraintTop_toTopOf="parent"/>
+
+                <TextView
+                    android:id="@+id/tv_coupon_count"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_marginTop="10dp"
+                    android:paddingHorizontal="18dp"
+                    android:paddingVertical="6dp"
+                    android:background="@drawable/shape_rect_side_round"
+                    android:backgroundTint="#E4D9F8"
+                    android:text="쿠모로 4개 모았어요!"
+                    android:textSize="12sp"
+                    android:textColor="#4E4E4E"
+                    android:fontFamily="@font/pretendard_600"
+                    app:layout_constraintStart_toStartOf="@id/tv_store_name"
+                    app:layout_constraintTop_toBottomOf="@id/tv_store_name"/>
+
+                <androidx.recyclerview.widget.RecyclerView
+                    android:id="@+id/rv_menu"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:layout_marginTop="40dp"
+                    android:paddingHorizontal="@dimen/horizontal_padding"
+                    app:layout_constraintTop_toBottomOf="@+id/tv_coupon_count"/>
+
+            </androidx.constraintlayout.widget.ConstraintLayout>
+
+        </androidx.core.widget.NestedScrollView>
+    </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
