--- app/src/main/res/layout/fragment_home_sub.xml
@@ -1,52 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <data>
-
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:id="@+id/section_tool_bar"
-            android:layout_width="match_parent"
-            style="@style/TabBarDefaultStyle"
-            app:layout_constraintTop_toTopOf="parent">
-
-            <androidx.constraintlayout.utils.widget.ImageFilterView
-                android:id="@+id/btn_toolbar_back"
-                android:layout_width="30dp"
-                android:layout_height="30dp"
-                android:src="@drawable/icon_chevron_left"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent"
-                app:layout_constraintBottom_toBottomOf="parent"/>
-
-            <TextView
-                android:id="@+id/tv_toolBar_title"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="카페/디저트"
-                android:textColor="@color/main"
-                android:textSize="18sp"
-                android:fontFamily="@font/pretendard_600"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintTop_toTopOf="parent"
-                app:layout_constraintBottom_toBottomOf="parent"/>
-
-        </androidx.constraintlayout.widget.ConstraintLayout>
-
-        <androidx.fragment.app.FragmentContainerView
-            android:id="@+id/fl_home_sub"
-            android:name="androidx.navigation.fragment.NavHostFragment"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            app:layout_constraintTop_toBottomOf="@id/section_tool_bar"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:navGraph="@navigation/nav_home_sub"/>
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
