+++ app/src/main/res/navigation/nav_home.xml
@@ -9,12 +9,30 @@
         android:name="com.umc.coumo.presentation.fragment.HomeMainFragment"
         android:label="HomeMainFragment" >
         <action
-            android:id="@+id/action_homeMainFragment_to_homeSubFragment"
-            app:destination="@id/homeSubFragment" />
-
+            android:id="@+id/action_homeMainFragment_to_homeListFragment"
+            app:destination="@id/homeListFragment" />
+        <action
+            android:id="@+id/action_homeMainFragment_to_homeDetailFragment"
+            app:destination="@id/homeDetailFragment" />
     </fragment>
     <fragment
-        android:id="@+id/homeSubFragment"
-        android:name="com.umc.coumo.presentation.fragment.HomeSubFragment"
-        android:label="HomeSubFragment" />
+        android:id="@+id/homeDetailFragment"
+        android:name="com.umc.coumo.presentation.fragment.HomeDetailFragment"
+        android:label="HomeDetailFragment" >
+        <action
+            android:id="@+id/action_homeDetailFragment_to_homeMenuFragment"
+            app:destination="@id/homeMenuFragment2" />
+    </fragment>
+    <fragment
+        android:id="@+id/homeListFragment"
+        android:name="com.umc.coumo.presentation.fragment.HomeListFragment"
+        android:label="HomeListFragment" >
+        <action
+            android:id="@+id/action_homeListFragment_to_homeDetailFragment"
+            app:destination="@id/homeDetailFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/homeMenuFragment2"
+        android:name="com.umc.coumo.presentation.fragment.HomeMenuFragment"
+        android:label="HomeMenuFragment" />
 </navigation>
